:truck: (imu): Rename C++ headers with .hpp (instead of .h)
