+++ libs/IMUKit/include/internal/Mahony.cpp
@@ -3,7 +3,7 @@
 // Copyright 2022 APF France handicap
 // SPDX-License-Identifier: Apache-2.0
 
-#include "Mahony.h"
+#include "Mahony.hpp"
 #include <cmath>
 #include <numbers>
