+++ libs/IMUKit/source/IMUKit.cpp
@@ -2,7 +2,7 @@
 // Copyright 2022 APF France handicap
 // SPDX-License-Identifier: Apache-2.0
 
-#include "IMUKit.h"
+#include "IMUKit.hpp"
 
 using namespace leka;
