+++ libs/MotionKit/include/MotionKit.hpp
@@ -7,8 +7,8 @@
 #include <interface/drivers/Motor.h>
 #include <interface/libs/EventLoop.h>
 
-#include "IMUKit.h"
-#include "PID.h"
+#include "IMUKit.hpp"
+#include "PID.hpp"
 #include "interface/drivers/Timeout.h"
 
 namespace leka {
