+++ libs/MotionKit/source/MotionKit.cpp
@@ -2,7 +2,7 @@
 // Copyright 2022 APF France handicap
 // SPDX-License-Identifier: Apache-2.0
 
-#include "MotionKit.h"
+#include "MotionKit.hpp"
 
 #include "MathUtils.h"
 #include "ThisThread.h"
