+++ spikes/lk_reinforcer/main.cpp
@@ -18,7 +18,7 @@
 #include "CoreLCD.hpp"
 #include "CoreLED.h"
 #include "CoreLL.h"
-#include "CoreLSM6DSOX.h"
+#include "CoreLSM6DSOX.hpp"
 #include "CoreLTDC.hpp"
 #include "CoreMotor.h"
 #include "CorePwm.h"
@@ -30,10 +30,10 @@
 #include "EventLoopKit.h"
 #include "FATFileSystem.h"
 #include "HelloWorld.h"
-#include "IMUKit.h"
+#include "IMUKit.hpp"
 #include "LedKit.h"
 #include "LogKit.h"
-#include "MotionKit.h"
+#include "MotionKit.hpp"
 #include "ReinforcerKit.h"
 #include "SDBlockDevice.h"
 #include "VideoKit.h"
